@@ -1,13 +1,13 @@
 cmake_minimum_required(VERSION 3.30.3)
 project(TradeZest VERSION 1.0.0 LANGUAGES CXX)
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 
 ### DEPENDENCIES ###
 
-# Boost
+# Boost (requires boost to be already installed in the workstation)
 find_package(Boost REQUIRED COMPONENTS ALL)
 
 include_directories(${Boost_INCLUDE_DIR})
@@ -75,6 +75,7 @@ add_custom_command(
 enable_testing()
 
 set(TEST_FILES
+    HTTPRequest
     OandaApi
 )
 
